tests/wpt/web-platform-tests/fullscreen/api/document-exit-fullscreen-manual.html

<!DOCTYPE html>
<title>Document#exitFullscreen()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<script>
async_test(function(t)
{
    trusted_request(document.querySelector("div"));

    document.addEventListener("fullscreenchange", t.step_func(function(event)
    {
        if (document.fullscreenElement) {
            document.exitFullscreen();
        } else {
            assert_equals(event.target, document, "event.target");
            assert_false(event.bubbles, "event.bubbles");
            assert_false(event.cancelable, "event.cancelable");
            t.done();
        }
    }));
});
</script>
